requirements.txt
pyinstaller==6.12.0
dearpygui==2.0.0
pillow==11.1.0
numpy==2.2.4